# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Contract Deployer API",
    "description": "Flask API for compiling and deploying smart contract projects using Docker containers with network support",
    "version": "1.0.0"
  },
...
  "paths": {
    "/deploy": {
      "post": {
        "tags": ["Deployment"],
        "summary": "Deploy and compile smart contract project",
        "description": "Compiles a Hardhat or Foundry project in a Docker container and returns artifacts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["path_url", "framework"],
                "properties": {
                  "path_url": {
                    "type": "string",
                    "description": "Local path to the smart contract project",
                    "example": "./cloned/eurodollar-protocol"
                  },
                  "framework": {
                    "type": "string",
                    "enum": ["hardhat", "foundry"],
                    "description": "Smart contract framework to use",
                    "example": "foundry"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compilation successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "artifact_path": {
                      "type": "string",
                      "example": "/path/to/project/.artifacts"
                    },
                    "stdout": {
                      "type": "string",
                      "description": "Last 2000 characters of stdout"
                    },
                    "stderr": {
                      "type": "string",
                      "description": "Last 2000 characters of stderr"
                    },
                    "duration_seconds": {
                      "type": "number",
                      "format": "float",
                      "example": 45.2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Missing path_url in request body"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Path not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Path does not exist: ./invalid/path"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Compilation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "artifact_path": {
                      "type": "string",
                      "example": ""
                    },
                    "stdout": {
                      "type": "string"
                    },
                    "stderr": {
                      "type": "string"
                    },
                    "duration_seconds": {
                      "type": "number",
                      "format": "float"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Docker unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Docker connection failed. Is Docker running?"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/publish": {
      "post": {
        "tags": ["Deployment"],
        "summary": "Deploy contracts on-chain",
        "description": "Deploys Foundry contracts on-chain using forge script and caches results with network support",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["path_url", "framework"],
                "properties": {
                  "path_url": {
                    "type": "string",
                    "description": "Local path to the Foundry project",
                    "example": "./cloned/eurodollar-protocol"
                  },
                  "framework": {
                    "type": "string",
                    "enum": ["foundry", "hardhat"],
                    "description": "Smart contract framework",
                    "example": "foundry"
                  },
                  "network": {
                    "type": "string",
                    "description": "Network to deploy to",
                    "enum": ["local", "ETH_TESTNET", "BASE_TESTNET", "BSC_TESTNET"],
                    "example": "local"
                  },
                  "script_path": {
                    "type": "string",
                    "description": "Path to the deployment script",
                    "example": "script/Deploy.s.sol"
                  },
                  "fork": {
                    "type": "boolean",
                    "description": "Use forked anvil for local networks (default: false)",
                    "example": false,
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deployment successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "deployed_contracts": {
                      "type": "object",
                      "description": "Map of contract names to deployed addresses",
                      "example": {
                        "Validator": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
                        "USDE": "0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0"
                      }
                    },
                    "deployment_key": {
                      "type": "string",
                      "description": "Unique key for this deployment",
                      "example": "./cloned/eurodollar-protocol:1692547200"
                    },
                    "stdout": {
                      "type": "string",
                      "description": "Last 2000 characters of stdout"
                    },
                    "stderr": {
                      "type": "string",
                      "description": "Last 2000 characters of stderr"
                    },
                    "duration_seconds": {
                      "type": "number",
                      "format": "float",
                      "example": 45.2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Missing framework in request body"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Path or script not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Deploy.s.sol script not found in script/ directory"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Deployment failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "deployed_contracts": {
                      "type": "object",
                      "example": {}
                    },
                    "stdout": {
                      "type": "string"
                    },
                    "stderr": {
                      "type": "string"
                    },
                    "duration_seconds": {
                      "type": "number",
                      "format": "float"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Docker unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Docker connection failed. Is Docker running?"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}